Moved loadModelFromResourceFolder to separate file
