+++ demo/shared/src/main/resources/games/demo/ship/main
@@ -1,0 +1,3 @@
+name=
+obj=
+mtl=
