+++ demo/shared/src/main/scala/games/demo/Engine.scala
@@ -68,96 +68,6 @@ class Engine(itf: EngineInterface, localEC: ExecutionContext, parEC: ExecutionCo
       conn.write(data)
   }
 
-  def loadModelFromResource(resourceFolder: String): Future[OpenGLMesh] = {
-    val mainResource = Resource(resourceFolder + "/main")
-    val mainFileFuture = Utils.getTextDataFromResource(mainResource)
-    val mainFuture = for (mainFile <- mainFileFuture) yield {
-      val mainLines = Utils.lines(mainFile)
-
-      var nameOpt: Option[String] = None
-      var objPathOpt: Option[String] = None
-      val mtlPaths: mutable.Queue[String] = mutable.Queue()
-
-      mainLines.foreach { line =>
-        val tokens = line.split("=", 2)
-        if (tokens.size != 2) throw new RuntimeException("Main model file malformed: \"" + line + "\"")
-        val key = tokens(0)
-        val value = tokens(1)
-
-        key match {
-          case "name" => nameOpt = Some(value)
-          case "obj"  => objPathOpt = Some(value)
-          case "mtl"  => mtlPaths += value
-          case _      => Console.err.println("Warning: unknown model key in line: \"" + line + "\"")
-        }
-      }
-
-      def missing(missingKey: String) = throw new RuntimeException("Missing key for " + missingKey + " in model")
-
-      val name = nameOpt.getOrElse(missing("name"))
-      val objPath = objPathOpt.getOrElse(missing("obj path"))
-
-      val objResource = Resource(resourceFolder + "/" + objPath)
-      val objFileFuture = Utils.getTextDataFromResource(objResource)
-
-      val mtlFileFutures = for (mtlPath <- mtlPaths) yield {
-        val mtlResource = Resource(resourceFolder + "/" + mtlPath)
-        val mtlFileFuture = Utils.getTextDataFromResource(mtlResource)
-        mtlFileFuture
-      }
-
-      val mtlFilesFuture = Future.sequence(mtlFileFutures)
-
-      for (
-        objFile <- objFileFuture;
-        mtlFiles <- mtlFilesFuture
-      ) yield {
-        val objLines = Utils.lines(objFile)
-        val mtlLines = mtlPaths.zip(mtlFiles.map(Utils.lines(_))).toMap
-
-        val objs = SimpleOBJParser.parseOBJ(objLines, mtlLines)
-        val meshes = SimpleOBJParser.convOBJObjectToTriMesh(objs)
-
-        val mesh = meshes(name)
-
-        val meshVerticesCount = mesh.vertices.length
-        val verticesData = GLES2.createFloatBuffer(meshVerticesCount * 3)
-        mesh.vertices.foreach { v => v.store(verticesData) }
-        verticesData.flip()
-        val verticesBuffer = gl.createBuffer()
-        gl.bindBuffer(GLES2.ARRAY_BUFFER, verticesBuffer)
-        gl.bufferData(GLES2.ARRAY_BUFFER, verticesData, GLES2.STATIC_DRAW)
-        val normals = mesh.normals.get
-        val normalsData = GLES2.createFloatBuffer(meshVerticesCount * 3); require(meshVerticesCount == normals.length)
-        normals.foreach { v => v.store(normalsData) }
-        normalsData.flip()
-        val normalsBuffer = gl.createBuffer()
-        gl.bindBuffer(GLES2.ARRAY_BUFFER, normalsBuffer)
-        gl.bufferData(GLES2.ARRAY_BUFFER, normalsData, GLES2.STATIC_DRAW)
-        val openGLSubMeshes = mesh.submeshes.map { submesh =>
-          val tris = submesh.tris
-          val submeshVerticesCount = tris.length * 3
-          val indicesData = GLES2.createShortBuffer(submeshVerticesCount)
-          tris.foreach {
-            case (i0, i1, i2) =>
-              indicesData.put(i0.toShort)
-              indicesData.put(i1.toShort)
-              indicesData.put(i2.toShort)
-          }
-          indicesData.flip()
-          val indicesBuffer = gl.createBuffer()
-          gl.bindBuffer(GLES2.ELEMENT_ARRAY_BUFFER, indicesBuffer)
-          gl.bufferData(GLES2.ELEMENT_ARRAY_BUFFER, indicesData, GLES2.STATIC_DRAW)
-          OpenGLSubMesh(indicesBuffer, submeshVerticesCount, submesh.material.get.ambientColor.get, submesh.material.get.diffuseColor.get)
-        }
-
-        OpenGLMesh(verticesBuffer, normalsBuffer, meshVerticesCount, openGLSubMeshes)
-      }
-    }
-
-    Utils.reduceFuture(mainFuture)
-  }
-
   def continue(): Boolean = continueCond
 
   def onClose(): Unit = {
